refactor: share mod context menu between grid cards and list rows
