+++ src/renderer/src/components/mod/mod-card.tsx
@@ -8,6 +8,7 @@ import { CalendarIcon, FolderIcon } from "lucide-react";
 import { memo, useRef } from "react";
 import { toast } from "sonner";
 import { ModCardHeader } from "./mod-card-header";
+import { ModContextMenu } from "./mod-context-menu";
 import { ModIniList } from "./mod-ini-list";
 import { ModPreviewContainer } from "./mod-preview-container";
 import { getModColorClass } from "./utils";
@@ -16,6 +17,16 @@ interface ModCardProps {
   mod: ModInfo;
   selectedGroupPath?: string;
   onToggle: (mod: ModInfo, event?: React.MouseEvent) => void;
+  fixTools: {
+    id: string;
+    name: string;
+    type: string;
+    size: number;
+  }[];
+  presets: {
+    id: string;
+    name: string;
+  }[];
   onToggleKeyUpdate: (
     modPath: string,
     iniPath: string,
@@ -29,6 +40,8 @@ export const ModCard = memo(function ModCard({
   mod,
   selectedGroupPath,
   onToggle,
+  fixTools,
+  presets,
   onToggleKeyUpdate,
 }: ModCardProps) {
   const { queryClient } = useRouteContext({ from: "__root__" });
@@ -100,65 +113,76 @@ export const ModCard = memo(function ModCard({
   };
 
   return (
-    <div
-      className={cn(
-        "rounded-sm overflow-hidden border-border/75 cursor-pointer p-1 h-[400px] relative hover:shadow-lg transition-shadow duration-150",
-        getModColorClass(mod.isEnabled),
-      )}
-      onMouseDown={(e) => {
-        mouseDownTargetRef.current = e.target;
-      }}
-      onClick={(e) => {
-        const target = mouseDownTargetRef.current as HTMLElement;
-        if (target && (target.tagName === "INPUT" || target.closest("button"))) {
-          return;
-        }
-        onToggle(mod, e);
-      }}
-      draggable={false}
+    <ModContextMenu
+      mod={mod}
+      selectedGroupPath={selectedGroupPath}
+      fixTools={fixTools}
+      presets={presets}
     >
-      {mod.preview?.match(/\.(jpeg|jpg|gif|png|webp|bmp)$/i) && (
-        <div
-          className="absolute inset-0 z-0 blur-lg scale-110 pointer-events-none opacity-25"
-          style={{ transform: "translateZ(0)", willChange: "filter" }}
-        >
-          <img src={`local://${mod.preview}`} alt="preview" className="w-full h-full object-fill" />
-        </div>
-      )}
-
-      <ModCardHeader mod={mod} selectedGroupPath={selectedGroupPath} />
-
-      <div className="flex flex-row h-[calc(100%-2rem)] space-x-2 relative z-10">
-        <ModPreviewContainer
-          mod={mod}
-          selectedGroupPath={selectedGroupPath}
-          onPaste={handlePaste}
-        />
-
-        {mod.inis.length > 0 && (
-          <>
-            <Separator orientation="vertical" />
-            <ModIniList mod={mod} onToggleKeyUpdate={onToggleKeyUpdate} />
-          </>
+      <div
+        className={cn(
+          "rounded-sm overflow-hidden border-border/75 cursor-pointer p-1 h-[400px] relative hover:shadow-lg transition-shadow duration-150",
+          getModColorClass(mod.isEnabled),
         )}
-      </div>
+        onMouseDown={(e) => {
+          mouseDownTargetRef.current = e.target;
+        }}
+        onClick={(e) => {
+          const target = mouseDownTargetRef.current as HTMLElement;
+          if (target && (target.tagName === "INPUT" || target.closest("button"))) {
+            return;
+          }
+          onToggle(mod, e);
+        }}
+        draggable={false}
+      >
+        {mod.preview?.match(/\.(jpeg|jpg|gif|png|webp|bmp)$/i) && (
+          <div
+            className="absolute inset-0 z-0 blur-lg scale-110 pointer-events-none opacity-25"
+            style={{ transform: "translateZ(0)", willChange: "filter" }}
+          >
+            <img
+              src={`local://${mod.preview}`}
+              alt="preview"
+              className="w-full h-full object-fill"
+            />
+          </div>
+        )}
 
-      <div className="absolute left-1 bottom-1 flex flex-col space-y-1 z-10">
-        <Badge
-          className="bg-background/35 backdrop-blur text-foreground text-xs h-5 flex items-center gap-1.5"
-          style={{ transform: "translateZ(0)", willChange: "backdrop-filter" }}
-        >
-          <FolderIcon />
-          {formatSize(mod.size)}
-        </Badge>
-        <Badge
-          className="bg-background/35 backdrop-blur text-foreground text-xs h-5 flex items-center gap-1.5"
-          style={{ transform: "translateZ(0)", willChange: "backdrop-filter" }}
-        >
-          <CalendarIcon />
-          {formatDate(new Date(mod.mtime), "ko")}
-        </Badge>
+        <ModCardHeader mod={mod} selectedGroupPath={selectedGroupPath} />
+
+        <div className="flex flex-row h-[calc(100%-2rem)] space-x-2 relative z-10">
+          <ModPreviewContainer
+            mod={mod}
+            selectedGroupPath={selectedGroupPath}
+            onPaste={handlePaste}
+          />
+
+          {mod.inis.length > 0 && (
+            <>
+              <Separator orientation="vertical" />
+              <ModIniList mod={mod} onToggleKeyUpdate={onToggleKeyUpdate} />
+            </>
+          )}
+        </div>
+
+        <div className="absolute left-1 bottom-1 flex flex-col space-y-1 z-10">
+          <Badge
+            className="bg-background/35 backdrop-blur text-foreground text-xs h-5 flex items-center gap-1.5"
+            style={{ transform: "translateZ(0)", willChange: "backdrop-filter" }}
+          >
+            <FolderIcon />
+            {formatSize(mod.size)}
+          </Badge>
+          <Badge
+            className="bg-background/35 backdrop-blur text-foreground text-xs h-5 flex items-center gap-1.5"
+            style={{ transform: "translateZ(0)", willChange: "backdrop-filter" }}
+          >
+            <CalendarIcon />
+            {formatDate(new Date(mod.mtime), "ko")}
+          </Badge>
+        </div>
       </div>
-    </div>
+    </ModContextMenu>
   );
 });
