+++ src/renderer/src/components/mod/mod-grid.tsx
@@ -2,6 +2,7 @@ import { ScrollArea } from "@renderer/components/ui/scroll-area";
 import { Skeleton } from "@renderer/components/ui/skeleton";
 import { useDelayedSkeleton } from "@renderer/hooks/use-delayed-skeleton";
 import { useFilteredMods } from "@renderer/hooks/use-filtered-mods";
+import { useModContextMenuData } from "@renderer/hooks/use-mod-context-menu-data";
 import { useModGroup } from "@renderer/hooks/use-mod-data";
 import { useModMutations } from "@renderer/hooks/use-mod-mutations";
 import { useVirtualizationSettings } from "@renderer/hooks/use-settings";
@@ -29,6 +30,7 @@ export function ModGrid(_props: ModGridProps) {
 
   const { toggleModMutation, exclusiveToggleModMutation, updateToggleKeyMutation } =
     useModMutations();
+  const { fixTools, presets } = useModContextMenuData();
 
   const mods = useFilteredMods(activeGroup?.mods || [], searchQuery);
   const isLoading = isPending || isPlaceholderData;
@@ -165,6 +167,8 @@ export function ModGrid(_props: ModGridProps) {
                   mod={mod}
                   selectedGroupPath={selectedGroupPath}
                   onToggle={handleToggle}
+                  fixTools={fixTools}
+                  presets={presets}
                   onToggleKeyUpdate={handleToggleKeyUpdate}
                 />
               ))}
@@ -207,6 +211,8 @@ export function ModGrid(_props: ModGridProps) {
                           mod={mod}
                           selectedGroupPath={selectedGroupPath}
                           onToggle={handleToggle}
+                          fixTools={fixTools}
+                          presets={presets}
                           onToggleKeyUpdate={handleToggleKeyUpdate}
                         />
                       ))}
