+++ src/renderer/src/components/mod/mod-list.tsx
@@ -2,11 +2,12 @@ import { ScrollArea } from "@renderer/components/ui/scroll-area";
 import { Skeleton } from "@renderer/components/ui/skeleton";
 import { useDelayedSkeleton } from "@renderer/hooks/use-delayed-skeleton";
 import { useFilteredMods } from "@renderer/hooks/use-filtered-mods";
+import { useModContextMenuData } from "@renderer/hooks/use-mod-context-menu-data";
 import { useModGroup } from "@renderer/hooks/use-mod-data";
 import { useModMutations } from "@renderer/hooks/use-mod-mutations";
 import { useModStore } from "@renderer/store/mod";
 import type { ModInfo } from "@renderer/types/mod";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback } from "react";
 import { useTranslation } from "react-i18next";
 import { ListHead } from "./mod-list-head";
 import { ModListRow } from "./mod-list-row";
@@ -23,20 +24,11 @@ export function ModList(_props: ModListProps) {
   const { data: activeGroup, isPlaceholderData, isPending } = useModGroup(selectedGroupPath);
 
   const { toggleModMutation, exclusiveToggleModMutation } = useModMutations();
+  const { fixTools, presets } = useModContextMenuData();
 
   const mods = useFilteredMods(activeGroup?.mods || [], searchQuery);
   const isLoading = isPending || isPlaceholderData;
   const showSkeleton = useDelayedSkeleton(isLoading);
-
-  const [fixTools, setFixTools] = useState<
-    { id: string; name: string; type: string; size: number }[]
-  >([]);
-  const [presets, setPresets] = useState<{ id: string; name: string }[]>([]);
-
-  useEffect(() => {
-    window.api.invoke("ftm:getScripts").then((res) => setFixTools(res || []));
-    window.api.invoke("ftm:getPresets").then((res) => setPresets(res || []));
-  }, []);
 
   const handleToggle = useCallback(
     (mod: ModInfo, event?: React.MouseEvent) => {
